// src/common/models/types/time.ts
export type Time =
  | 'milliseconds'
  | 'millisecond'
  | 'ms'
  | 'seconds'
  | 'second'
  | 's'
  | 'minutes'
  | 'minute'
  | 'm'
  | 'hours'
  | 'hour'
  | 'h'
  | 'days'
  | 'day'
  | 'd'
  | 'weeks'
  | 'week'
  | 'w'
  | 'months'
  | 'month'
  | 'M'
  | 'quarters'
  | 'quarter'
  | 'q'
  | 'years'
  | 'year'
  | 'y';